Cast the flags passed to sws_getContext to a 16-bit int. Otherwise I get segfaults on 64-bit linux.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@958 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ libhb/hb.c
@@ -369,7 +369,7 @@ void hb_get_preview( hb_handle_t * h, hb_title_t * title, int picture,
                              title->height - (job->crop[0] + job->crop[1]),
                              PIX_FMT_YUV420P,
                              job->width, job->height, PIX_FMT_YUV420P,
-                             SWS_LANCZOS|SWS_ACCURATE_RND, NULL, NULL, NULL);
+                             (uint16_t)(SWS_LANCZOS|SWS_ACCURATE_RND), NULL, NULL, NULL);
 
     // Scale
     sws_scale(context,
@@ -383,7 +383,7 @@ void hb_get_preview( hb_handle_t * h, hb_title_t * title, int picture,
     // Get preview context
     context = sws_getContext(job->width, job->height, PIX_FMT_YUV420P,
                              job->width, job->height, PIX_FMT_RGBA32,
-                             SWS_LANCZOS|SWS_ACCURATE_RND, NULL, NULL, NULL);
+                             (uint16_t)(SWS_LANCZOS|SWS_ACCURATE_RND), NULL, NULL, NULL);
 
     // Create preview
     sws_scale(context,
